Fixed x/auth genesis state initialization inside sim tests
